Deeper Recursion
Created a module that iterated through the "Decrease and Conquer"
approach to recursion by using a factorial example.

A more complex module was built to demonstrate the "Divide and
Conquer" approach to recursion. This module starts with a list,
divides it until it's left with a group of single element lists,
and then compares them to merge back into a single list that's
sorted ascending